+++ release/src/router/jitterentropy-rngd/jitterentropy-base-user.h
@@ -1,7 +1,7 @@
 /*
  * Non-physical true random number generator based on timing jitter.
  *
- * Copyright Stephan Mueller <smueller@chronox.de>, 2013 - 2021
+ * Copyright Stephan Mueller <smueller@chronox.de>, 2013 - 2022
  *
  * License
  * =======
@@ -44,19 +44,15 @@
 
 /*
  * Set the following defines as needed for your environment
+ * Compilation for AWS-LC     #define AWSLC
+ * Compilation for libgcrypt  #define LIBGCRYPT
+ * Compilation for OpenSSL    #define OPENSSL
  */
-/* Compilation for libgcrypt */
-#ifndef LIBGCRYPT
-#undef LIBGCRYPT
-#endif
 
-/* Compilation for OpenSSL */
-#ifndef OPENSSL
-#undef OPENSSL
-#endif
-
+#include <limits.h>
 #include <time.h>
 #include <stdint.h>
+#include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
 #include <sys/types.h>
@@ -65,6 +61,7 @@
 #include <fcntl.h>
 #include <unistd.h>
 #include <errno.h>
+#include <sched.h>
 
 /* Timer-less entropy source */
 #ifdef JENT_CONF_ENABLE_INTERNAL_TIMER
@@ -83,6 +80,10 @@
 #endif
 #endif
 
+#if defined(AWSLC)
+#include <openssl/crypto.h>
+#endif
+
 #ifdef __MACH__
 #include <assert.h>
 #include <CoreServices/CoreServices.h>
@@ -91,11 +92,18 @@
 #include <unistd.h>
 #endif
 
-#ifdef __x86_64__
+#if (__x86_64__) || (__i386__)
+/* Support rdtsc read on 64-bit and 32-bit x86 architectures */
 
+#ifdef __x86_64__
 # define DECLARE_ARGS(val, low, high)    unsigned long low, high
 # define EAX_EDX_VAL(val, low, high)     ((low) | (high) << 32)
 # define EAX_EDX_RET(val, low, high)     "=a" (low), "=d" (high)
+#elif __i386__
+# define DECLARE_ARGS(val, low, high)    unsigned long val
+# define EAX_EDX_VAL(val, low, high)     val
+# define EAX_EDX_RET(val, low, high)     "=A" (val)
+#endif
 
 static inline void jent_get_nstime(uint64_t *out)
 {
@@ -104,7 +112,7 @@ static inline void jent_get_nstime(uint64_t *out)
 	*out = EAX_EDX_VAL(val, low, high);
 }
 
-#else /* __x86_64__ */
+#else /* (__x86_64__) || (__i386__) */
 
 static inline void jent_get_nstime(uint64_t *out)
 {
@@ -151,7 +159,7 @@ static inline void *jent_zalloc(size_t len)
 	 * decision for less memory protection. */
 #define CONFIG_CRYPTO_CPU_JITTERENTROPY_SECURE_MEMORY
 	tmp = gcry_xmalloc_secure(len);
-#elif defined(OPENSSL)
+#elif defined(OPENSSL) || defined(AWSLC)
 	/* Does this allocation implies secure memory use? */
 	tmp = OPENSSL_malloc(len);
 #else
@@ -169,6 +177,10 @@ static inline void jent_zfree(void *ptr, unsigned int len)
 #ifdef LIBGCRYPT
 	memset(ptr, 0, len);
 	gcry_free(ptr);
+#elif defined(AWSLC)
+    /* AWS-LC stores the length of allocated memory internally and automatically wipes it in OPENSSL_free */
+	(void) len;
+	OPENSSL_free(ptr);
 #elif defined(OPENSSL)
 	OPENSSL_cleanse(ptr, len);
 	OPENSSL_free(ptr);
@@ -182,6 +194,8 @@ static inline int jent_fips_enabled(void)
 {
 #ifdef LIBGCRYPT
 	return fips_mode();
+#elif defined(AWSLC)
+	return FIPS_mode();
 #elif defined(OPENSSL)
 #ifdef OPENSSL_FIPS
 	return FIPS_mode();
@@ -206,8 +220,173 @@ static inline int jent_fips_enabled(void)
 
 static inline void jent_memset_secure(void *s, size_t n)
 {
+#if defined(AWSLC)
+	OPENSSL_cleanse(s, n);
+#else
 	memset(s, 0, n);
 	__asm__ __volatile__("" : : "r" (s) : "memory");
+#endif
+}
+
+static inline long jent_ncpu(void)
+{
+#ifdef _POSIX_SOURCE
+	long ncpu = sysconf(_SC_NPROCESSORS_ONLN);
+
+	if (ncpu == -1)
+		return -errno;
+
+	if (ncpu == 0)
+		return -EFAULT;
+
+	return ncpu;
+#else
+	return 1;
+#endif
+}
+
+#ifdef __linux__
+
+# if defined(_SC_LEVEL1_DCACHE_SIZE) &&					\
+     defined(_SC_LEVEL2_CACHE_SIZE) &&					\
+     defined(_SC_LEVEL3_CACHE_SIZE)
+
+static inline void jent_get_cachesize(long *l1, long *l2, long *l3)
+{
+	*l1 = sysconf(_SC_LEVEL1_DCACHE_SIZE);
+	*l2 = sysconf(_SC_LEVEL2_CACHE_SIZE);
+	*l3 = sysconf(_SC_LEVEL3_CACHE_SIZE);
+}
+
+# else
+
+static inline void jent_get_cachesize(long *l1, long *l2, long *l3)
+{
+#define JENT_SYSFS_CACHE_DIR "/sys/devices/system/cpu/cpu0/cache"
+	long val;
+	unsigned int i;
+	char buf[10], file[50];
+	int fd = 0;
+
+	/* Iterate over all caches */
+	for (i = 0; i < 4; i++) {
+		unsigned int shift = 0;
+		char *ext;
+
+		/*
+		 * Check the cache type - we are only interested in Unified
+		 * and Data caches.
+		 */
+		memset(buf, 0, sizeof(buf));
+		snprintf(file, sizeof(file), "%s/index%u/type",
+			 JENT_SYSFS_CACHE_DIR, i);
+		fd = open(file, O_RDONLY);
+		if (fd < 0)
+			continue;
+		while (read(fd, buf, sizeof(buf)) < 0 && errno == EINTR);
+		close(fd);
+		buf[sizeof(buf) - 1] = '\0';
+
+		if (strncmp(buf, "Data", 4) && strncmp(buf, "Unified", 7))
+			continue;
+
+		/* Get size of cache */
+		memset(buf, 0, sizeof(buf));
+		snprintf(file, sizeof(file), "%s/index%u/size",
+			 JENT_SYSFS_CACHE_DIR, i);
+
+		fd = open(file, O_RDONLY);
+		if (fd < 0)
+			continue;
+		while (read(fd, buf, sizeof(buf)) < 0 && errno == EINTR);
+		close(fd);
+		buf[sizeof(buf) - 1] = '\0';
+
+		ext = strstr(buf, "K");
+		if (ext) {
+			shift = 10;
+			*ext = '\0';
+		} else {
+			ext = strstr(buf, "M");
+			if (ext) {
+				shift = 20;
+				*ext = '\0';
+			}
+		}
+
+		val = strtol(buf, NULL, 10);
+		if (val == LONG_MAX)
+			continue;
+		val <<= shift;
+
+		if (!*l1)
+			*l1 = val;
+		else if (!*l2)
+			*l2 = val;
+		else {
+			*l3 = val;
+			break;
+		}
+	}
+#undef JENT_SYSFS_CACHE_DIR
+}
+
+# endif
+
+static inline uint32_t jent_cache_size_roundup(void)
+{
+	static int checked = 0;
+	static uint32_t cache_size = 0;
+
+	if (!checked) {
+		long l1 = 0, l2 = 0, l3 = 0;
+
+		jent_get_cachesize(&l1, &l2, &l3);
+		checked = 1;
+
+		/* Cache size reported by system */
+		if (l1 > 0)
+			cache_size += (uint32_t)l1;
+		if (l2 > 0)
+			cache_size += (uint32_t)l2;
+		if (l3 > 0)
+			cache_size += (uint32_t)l3;
+
+		/*
+		 * Force the output_size to be of the form
+		 * (bounding_power_of_2 - 1).
+		 */
+		cache_size |= (cache_size >> 1);
+		cache_size |= (cache_size >> 2);
+		cache_size |= (cache_size >> 4);
+		cache_size |= (cache_size >> 8);
+		cache_size |= (cache_size >> 16);
+
+		if (cache_size == 0)
+			return 0;
+
+		/*
+		 * Make the output_size the smallest power of 2 strictly
+		 * greater than cache_size.
+		 */
+		cache_size++;
+	}
+
+	return cache_size;
+}
+
+#else /* __linux__ */
+
+static inline uint32_t jent_cache_size_roundup(void)
+{
+	return 0;
+}
+
+#endif /* __linux__ */
+
+static inline void jent_yield(void)
+{
+	sched_yield();
 }
 
 /* --- helpers needed in user space -- */